+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Molio"
+        android:theme="@style/Theme.Molia"
         tools:targetApi="31" >
         <activity
             android:name=".ui.SaveActivity"
@@ -19,7 +19,7 @@
         <activity
             android:name=".ui.MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.Molio" >
+            android:theme="@style/Theme.Molia" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
