+++ app/src/main/java/com/hobarb/molia/SecondFragment.kt
@@ -1,4 +1,4 @@
-package com.hobarb.molio
+package com.hobarb.molia
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
-import com.hobarb.molio.databinding.FragmentSecondBinding
+import com.hobarb.molia.databinding.FragmentSecondBinding
 
 /**
  * A simple [Fragment] subclass as the second destination in the navigation.
