+++ app/src/main/java/com/hobarb/molia/interfaces/OnItemClickListener.interface.kt
@@ -1,4 +1,4 @@
-package com.hobarb.molio.interfaces
+package com.hobarb.molia.interfaces
 
 interface OnItemClickListener<T> {
     fun onItemClick(item: T)
