+++ app/src/main/java/com/hobarb/molia/models/dtos/SaveTitle.dto.kt
@@ -1,4 +1,4 @@
-package com.hobarb.molio.models.dtos
+package com.hobarb.molia.models.dtos
 
 import TitleDetails
 import java.io.Serializable
