+++ app/src/main/java/com/hobarb/molia/models/schemas/TitleDetails.model.kt
