+++ app/src/main/java/com/hobarb/molia/network/RetrofitInstance.kt
@@ -1,6 +1,6 @@
-package com.hobarb.molio.network
+package com.hobarb.molia.network
 
-import MolioBackendService
+import MoliaBackendService
 import OmdbApiService
 import com.google.gson.GsonBuilder
 import retrofit2.Retrofit
@@ -14,13 +14,13 @@ object RetrofitClient {
    // private const val BASE_URL_OMDB = "https://omdb.onrender.com"
    private const val BASE_URL_OMDB = "https://www.omdbapi.com"
 
-    val moliaApiService: MolioBackendService by lazy {
+    val moliaApiService: MoliaBackendService by lazy {
         val retrofit = Retrofit.Builder()
             .baseUrl(BASE_URL_MOLIA)
             .addConverterFactory(GsonConverterFactory.create())
             .build()
 
-        retrofit.create(MolioBackendService::class.java)
+        retrofit.create(MoliaBackendService::class.java)
     }
 
     val omdbApiService: OmdbApiService by lazy {
