+++ app/src/main/java/com/hobarb/molia/services/ApiHandler.kt
@@ -1,17 +1,12 @@
-package com.hobarb.molio.services
+package com.hobarb.molia.services
 
-import android.widget.Toast
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
-import com.hobarb.molio.models.dtos.SaveTitleModel
-import com.hobarb.molio.models.schemas.SearchTitle
-import com.hobarb.molio.network.RetrofitClient
-import org.json.JSONObject
+import com.hobarb.molia.models.dtos.SaveTitleModel
+import com.hobarb.molia.network.RetrofitClient
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.await
-import retrofit2.http.Query
 
 class ApiHandler {
