+++ app/src/main/res/values-v23/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
 
-    <style name="Theme.Molio" parent="Base.Theme.Molio">
+    <style name="Theme.Molia" parent="Base.Theme.Molia">
         <!-- Transparent system bars for edge-to-edge. -->
         <item name="android:navigationBarColor">@android:color/transparent</item>
         <item name="android:statusBarColor">@android:color/transparent</item>
